requirements.txt
Flask==3.0.0
PyGithub==2.1.1
ollama==0.1.6
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
markdown==3.5.1
